BAT-034: stabilize auth/session redirects and protected route guards
